📁 Getdata plugin gets a folder As it should
