+++ api/plugins/getData/error.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Error</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" type="text/css" href="index.css">
+    <link rel="stylesheet" type="text/css" href="data/index.css">
+</head>
+<body>
+    <h1 style="text-align: center;">Error 😓</h1>
+    <br>
+    <p>There was an error processing your request.</p>
+    <p>This problem was likely caused by the sensor server being down...</p>
+    <p>If the problem persists, please contact me at <a href="mailto:water@connorcode.com?subject=Error with Get Data Site&body={ERROR}">water@connorcode.com</a>.</p>
+    <br>
+    <p><strong>ERROR INFO:</strong> <code>{ERROR}</code></p>
+    <div class="footer">
+        <p class="footer-content">©Connor Slade · 2021 · V0.2 · <a href="https://github.com/Basicprogrammer10/WaterTemp">Github</a> · <a href="/">Home</a></p>
+    </div>
+</body>
+</html>
