+++ api/plugins/getData/index.scss
@@ -1,0 +1,18 @@
+body * {
+    text-align: center;
+    font-family: Arial,sans-serif;
+    margin: 0;
+    margin-top: 0;
+}
+
+a {
+    color: #58a6ff;
+}
+
+code {
+    padding: 2px 4px;
+    font-size: 90%;
+    color: #c7254e;
+    background-color: #f9f2f4;
+    border-radius: 4px;
+}
